+++ doc/source/changes.rst
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+0.4.0 - Future release
+----------------------
+* Add CLI tool for using the Warthog Client. See :doc:`cli`.
+* Add :meth:`warthog.client.WarthogClient.get_connections` method for getting the
+  number of active connections to a server.
+
 0.3.1 - 2014-10-17
 ------------------
 * Change ``setup.py`` script to not require setuptools.
@@ -8,13 +14,13 @@ Change Log
 0.3.0 - 2014-10-16
 ------------------
 * Add :doc:`install` documentation.
-* Change authentication request (:class:`warthog.api.SessionStartCommand`) to use ``POST``
+* Change authentication request (:class:`warthog.core.SessionStartCommand`) to use ``POST``
   instead of ``GET`` though there doesn't seem to be any actual difference as far as the
   load balancer API is concerned.
 
 0.2.0 - 2014-10-14
 ------------------
-* Add :doc:`design`, :doc:`usage`, and :doc:`api` documentation.
+* Add :doc:`design`, :doc:`usage`, and :doc:`lib` documentation.
 * Add test to ensure exported API is consistent.
 
 0.1.0 - 2014-10-11